+++ views/account.handlebars
@@ -1,1 +1,4 @@
 <h1>Account testing by Nathan</h1>
+
+<script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
+<script src="{{accountScript}}"></script>
