+++ views/login.handlebars
@@ -8,3 +8,5 @@
 </form>
 
 <a href="/signup" >Signup</a>
+<script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
+<script src="{{loginScript}}"></script>
